+++ CHANGELOG.md
@@ -1,14 +1,37 @@
 Changelog
 =========
+###v 0.8.3
+- Grunt configured
+- JSHint of code
+- Karma Grunt integration
+- Removed unused warning from tests
+- Test completed
+
+###v 0.8.2
+- Added some tests
+
+###v 0.8.0
+- Fixed return value for all functions
+- Added `destroy` method to clean all listeners
+
+###v 0.7.1
+- Correct return value in "run" method
+- Introduced a test suite
+
+###v 0.7.0
+- Request object has `get` method to retrieve parameters wherever they are passed
+- Bugfix: Splats do not contain query anymore in some cases
+
+
 ###v 0.6.3
-Added support for options. `ignorecase` option, default true
+- Added support for options. `ignorecase` option, default true
 
 ###v 0.6.2
-Added map file
-Bugfix: Query string do not appear in regular params value
-Internal : Added closure compiler support
+- Added map file
+- Bugfix: Query string do not appear in regular params value
+- Internal : Added closure compiler support
 
 ###v 0.6.1
-Added alias name for method `addRoute`. Now can be called `add`, `route`, `get`.
+- Added alias name for method `addRoute`. Now can be called `add`, `route`, `get`.
 
 Old changelog missing